+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ENHANCEMENTS:
 BUG FIXES:
 * alb: fix behavior of attribute `weight` in `http_backend` and `grpc_backend` and `stream_backend` in `yandex_alb_backend_group` resource
 * datatransfer: make `yandex_datatransfer_transfer` importable in accordance with the documentation
+* mdb: fix the optional parameter `preallocate` for kafka topic config when it is absent
 
 ## 0.70.0 (January 17, 2022)
 FEATURES: